+++ src/steps/os/unix.rs
@@ -643,26 +643,6 @@ pub fn run_pyenv(ctx: &ExecutionContext) -> Result<()> {
     ctx.run_type().execute(pyenv).arg("update").status_checked()
 }
 
-pub fn run_uv(ctx: &ExecutionContext) -> Result<()> {
-    let uv_exec = require("uv")?;
-    print_separator("uv");
-
-    ctx.run_type()
-        .execute(&uv_exec)
-        .args(["self", "update"])
-        .status_checked()
-        .ok();
-
-    // ignoring self-update errors, because they are likely due to uv's
-    // installation being managed by another package manager, in which
-    // case another step will handle the update.
-
-    ctx.run_type()
-        .execute(&uv_exec)
-        .args(["tool", "upgrade", "--all"])
-        .status_checked()
-}
-
 pub fn run_sdkman(ctx: &ExecutionContext) -> Result<()> {
     let bash = require("bash")?;
